requirements.txt
Click
netCDF4
python-dateutil
cf_units
jinja2
